Raspberry Pi specific hardware section
Just an idea; this isn't quite as fleshed out as I think a section like this could be, so far it's just a list and explanations, not how it goes together, which is very simple, but still bears mentioning. The Raspberry Pi is not exactly plug and play, certainly less so than Cubox, so for very first time users, I think a section like this could replace simple though tedious internet research into how to set up a Pi. Anything to make it easier for new folks.

I used Amazon store stock photos for the images, save for the IQaudIO board, which is off their site. Putting brand name items in a feed seems like advertising to me, which I personally don't like. (Except for the IOaudIO board; I know that Gordon helps alot on the site, so it would be good reciprocation. My two cents. And when I am stumping for RuneAudio.) Maybe a final draft would include black silhouettes of the items, as they are pretty generic in type. 

A simple vector image of the various hardware in relation to each other might be a good visual as well. I tend to be wordy, I learn best that way. The vector image would be for the visual learners. 

I am not in any way married to this concept or delivery. I got an idea and started writing because that's how I process. If there is a need/request for it later, we can put a revised draft in. For now, I leave it here for consideration.